# auth/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/auth/v1/authentication:
        post:
            tags:
                - Auth
            description: 认证接口
            operationId: Auth_Authentication
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.Response'
    /api/auth/v1/login:
        post:
            tags:
                - Auth
            description: 登录接口
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.Response'
    /api/auth/v1/logout:
        post:
            tags:
                - Auth
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.Response'
    /api/auth/v1/open_login:
        post:
            tags:
                - Auth
            operationId: Auth_OpenLogin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.Response'
    /api/auth/v1/open_login_info:
        get:
            tags:
                - Auth
            operationId: Auth_OpenLoginInfo
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.Response'
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
components:
    schemas:
        auth.v1.LoginRequest:
            type: object
            properties:
                UserName:
                    type: string
                PassWord:
                    type: string
            description: 登录请求
        auth.v1.OpenLoginResponse:
            type: object
            properties:
                Name:
                    type: string
                Icon:
                    type: string
                Href:
                    type: string
            description: 第三方登录响应
        auth.v1.Response:
            type: object
            properties:
                Code:
                    type: integer
                    format: uint32
                loginResponse:
                    $ref: '#/components/schemas/auth.v1.Response_LoginResponse'
                openLoginResponse:
                    $ref: '#/components/schemas/auth.v1.Response_OpenLoginResponses'
                Msg:
                    type: string
            description: 响应
        auth.v1.Response_LoginResponse:
            type: object
            properties:
                Token:
                    type: string
            description: |-
                oneof Data {
                    LoginResponse login_response = 2;
                    OpenLoginResponses open_login_response = 3;
                  }
                 登录响应
        auth.v1.Response_OpenLoginResponses:
            type: object
            properties:
                openLoginInfo:
                    type: array
                    items:
                        $ref: '#/components/schemas/auth.v1.OpenLoginResponse'
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Auth
    - name: Greeter
      description: The greeting service definition.